# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise AI Platform API
  version: 1.1.0
  description: >-
    Contratos HTTP do control plane e data plane. O tenant é derivado da
    identidade autenticada; clientes não podem escolher livremente outro tenant.
servers:
  - url: https://api.enterprise-ai-platform.example.com
    description: Referência de produção
  - url: http://localhost:8080
    description: Vertical slice local
security:
  - oidc:
      - agent.read
...
paths:
  /v1/agents:
    get:
      tags: [Agents]
      summary: Lista agentes visíveis para o tenant autenticado
      operationId: listAgents
      security:
        - oidc: [agent.read]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/PageToken'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AgentStatus'
      responses:
        '200':
          description: Página de agentes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Agents]
      summary: Registra uma nova versão de agente em DRAFT
      operationId: createAgent
      security:
        - oidc: [agent.write]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Versão criada
          headers:
            ETag:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/PolicyViolation'}
...
  /v1/agents/{agentId}:invoke:
    post:
      tags: [Invocation]
      summary: Invoca uma versão publicada do agente
      operationId: invokeAgent
      security:
        - oidc: [agent.invoke]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentInvokeRequest'
      responses:
        '200':
          description: Resposta completa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentInvokeResponse'
        '202':
          description: Execução assíncrona aceita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedExecution'
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/PolicyViolation'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/agents/{agentId}:stream:
    post:
      tags: [Invocation]
      summary: Invoca um agente com resposta via Server-Sent Events
      operationId: streamAgent
      security:
        - oidc: [agent.invoke]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentInvokeRequest'
      responses:
        '200':
          description: "Eventos SSE até event completed ou event failed"
          content:
            text/event-stream:
              schema:
                type: string
                examples:
                  - "event: token\ndata: {\"text\":\"Olá\"}\n\nevent: completed\ndata: {\"messageId\":\"msg-1\"}\n\n"
  /v1/agents/{agentId}:submit:
    post:
      tags: [Governance]
      summary: Submete uma versão para governança
      operationId: submitAgent
      security:
        - oidc: [governance.submit]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GovernanceSubmissionRequest'
      responses:
        '202':
          description: Submissão aceita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceDecision'
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/PolicyViolation'}
  /v1/agents/{agentId}:approve:
    post:
      tags: [Governance]
      summary: Aprova uma versão submetida
      operationId: approveAgent
      security:
        - oidc: [governance.approve]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GovernanceDecisionRequest'
      responses:
        '200':
          description: Decisão registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceDecision'
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/PolicyViolation'}
  /v1/agents/{agentId}:reject:
    post:
      tags: [Governance]
      summary: Rejeita uma versão submetida
      operationId: rejectAgent
      security:
        - oidc: [governance.approve]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GovernanceDecisionRequest'
      responses:
        '200':
          description: Decisão registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceDecision'
  /v1/agents/{agentId}:publish:
    post:
      tags: [Governance]
      summary: Publica uma versão previamente aprovada
      operationId: publishAgent
      security:
        - oidc: [agent.publish]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishAgentRequest'
      responses:
        '202':
          description: Publicação aceita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/PolicyViolation'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.com/.well-known/openid-configuration
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: UUID fornecido pelo cliente ou gerado pela plataforma.
      schema: {type: string, format: uuid}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      description: Obrigatório quando a invocação puder executar ferramentas com efeito colateral.
      schema: {type: string, minLength: 16, maxLength: 128}
    AgentId:
      name: agentId
      in: path
      required: true
      schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{2,63}$'}
...
    PageToken:
      name: pageToken
      in: query
      schema: {type: string}
    PageSize:
      name: pageSize
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: Identidade ausente ou inválida
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Forbidden:
      description: A identidade não possui autorização
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Recurso não encontrado ou não visível
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: Conflito de versão, estado ou idempotência
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    PolicyViolation:
      description: Requisição válida, porém bloqueada por política ou regra de negócio
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    TooManyRequests:
      description: Quota ou rate limit excedido
      headers:
        Retry-After:
          schema: {type: integer}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    InternalError:
      description: Erro interno sem exposição de detalhes sensíveis
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
    AgentStatus:
      type: string
      enum: [DRAFT, SUBMITTED, APPROVED, PUBLISHED, RETIRED]
    RiskClassification:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    Agent:
      type: object
      required: [agentId, name, version, owner, status, riskClassification, modelPolicy]
      properties:
        agentId: {type: string}
        name: {type: string}
        description: {type: string}
        version: {type: string}
        owner: {type: string}
        businessUnit: {type: string}
        status: {$ref: '#/components/schemas/AgentStatus'}
        riskClassification: {$ref: '#/components/schemas/RiskClassification'}
        modelPolicy: {$ref: '#/components/schemas/ModelPolicy'}
        allowedTools:
          type: array
          items: {type: string}
        knowledgeBaseIds:
          type: array
          items: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    CreateAgentRequest:
      type: object
      additionalProperties: false
      required: [agentId, name, version, owner, instructions, riskClassification, modelPolicy]
      properties:
        agentId: {type: string}
        name: {type: string}
        version: {type: string}
        owner: {type: string}
        businessUnit: {type: string}
        instructions: {type: string, minLength: 1}
        riskClassification: {$ref: '#/components/schemas/RiskClassification'}
        modelPolicy: {$ref: '#/components/schemas/ModelPolicy'}
        allowedTools:
          type: array
          items: {type: string}
        knowledgeBaseIds:
          type: array
          items: {type: string}
    ModelPolicy:
      type: object
      additionalProperties: false
      required: [capability, maxInputTokens, maxOutputTokens, temperature]
      properties:
        capability: {type: string, enum: [TEXT_GENERATION, EMBEDDING, MULTIMODAL]}
        preferredModels:
          type: array
          items: {type: string}
        maxInputTokens: {type: integer, minimum: 1}
        maxOutputTokens: {type: integer, minimum: 1}
        temperature: {type: number, minimum: 0, maximum: 2}
        maxCostUsd: {type: number, minimum: 0}
        allowedRegions:
          type: array
          items: {type: string}
    AgentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Agent'}
        nextPageToken: {type: [string, 'null']}
    AgentInvokeRequest:
      type: object
      additionalProperties: false
      required: [input, channel, sessionId]
      properties:
        input: {type: string, minLength: 1, maxLength: 50000}
        channel: {type: string}
        sessionId: {type: string}
        executionMode: {type: string, enum: [SYNC, ASYNC], default: SYNC}
        context:
          type: object
          additionalProperties: true
    AgentInvokeResponse:
      type: object
      required: [conversationId, messageId, answer, evaluationStatus, executionStatus]
      properties:
        conversationId: {type: string}
        messageId: {type: string}
        answer: {type: string}
        executionStatus: {type: string, enum: [SUCCESS, PARTIAL, FAILED, BLOCKED]}
        citations:
          type: array
          items: {$ref: '#/components/schemas/Citation'}
        toolCalls:
          type: array
          items: {$ref: '#/components/schemas/ToolCall'}
        tokenUsage: {$ref: '#/components/schemas/TokenUsage'}
        evaluationStatus: {type: string, enum: [QUEUED, APPROVED, WARNING, REJECTED, SKIPPED]}
    AcceptedExecution:
      type: object
      required: [executionId, status]
      properties:
        executionId: {type: string}
        status: {type: string, enum: [QUEUED, RUNNING]}
        statusUri: {type: string, format: uri}
    Citation:
      type: object
      required: [sourceId, title, chunkId, score]
      properties:
        sourceId: {type: string}
        title: {type: string}
        uri: {type: string}
        chunkId: {type: string}
        score: {type: number, minimum: 0, maximum: 1}
        dataClassification: {$ref: '#/components/schemas/DataClassification'}
    ToolCall:
      type: object
      required: [toolName, toolVersion, status]
      properties:
        toolName: {type: string}
        toolVersion: {type: string}
        status: {type: string, enum: [SUCCESS, FAILED, BLOCKED, PARTIAL]}
        latencyMs: {type: integer, minimum: 0}
        operationId: {type: string}
...
    GovernanceSubmissionRequest:
      type: object
      required: [agentVersion, riskClassification, evidence]
      properties:
        agentVersion: {type: string}
        riskClassification: {$ref: '#/components/schemas/RiskClassification'}
        evidence:
          type: array
          minItems: 1
          items: {type: string}
        notes: {type: string}
    GovernanceDecisionRequest:
      type: object
      required: [reason]
      properties:
        reason: {type: string, minLength: 10}
        conditions:
          type: array
          items: {type: string}
    GovernanceDecision:
      type: object
      required: [approvalId, agentId, decision, riskClassification]
      properties:
        approvalId: {type: string}
        agentId: {type: string}
        decision: {type: string, enum: [PENDING, APPROVED, REJECTED]}
        riskClassification: {$ref: '#/components/schemas/RiskClassification'}
        decidedBy: {type: [string, 'null']}
        decidedAt: {type: [string, 'null'], format: date-time}
    PublishAgentRequest:
      type: object
      required: [approvalId]
      properties:
        approvalId: {type: string}
        releaseNotes: {type: string}
...
    Problem:
      type: object
      required: [type, title, status, correlationId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string}
        correlationId: {type: string}
        violations:
          type: array
          items:
            type: object
            properties:
              field: {type: string}
              reason: {type: string}